+++ include/hdmi-cec-id.h
@@ -1,0 +1,48 @@
+#ifndef HDMI_IDS_H
+#define HDMI_IDS_H
+
+typedef enum {
+  CEC_ID_FEATURE_ABORT = 0x00,
+  CEC_ID_IMAGE_VIEW_ON = 0x04,
+  CEC_ID_TEXT_VIEW_ON = 0x0d,
+  CEC_ID_STANDBY = 0x36,
+  CEC_ID_USER_CONTROL_PRESSED = 0x44,
+  CEC_ID_USER_CONTROL_RELEASED = 0x45,
+  CEC_ID_GIVE_OSD_NAME = 0x46,
+  CEC_ID_SET_OSD_NAME = 0x47,
+  CEC_ID_SYSTEM_AUDIO_MODE_REQUEST = 0x70,
+  CEC_ID_GIVE_AUDIO_STATUS = 0x71,
+  CEC_ID_SET_SYSTEM_AUDIO_MODE = 0x72,
+  CEC_ID_GIVE_SYSTEM_AUDIO_MODE_STATUS = 0x7d,
+  CEC_ID_SYSTEM_AUDIO_MODE_STATUS = 0x7e,
+  CEC_ID_REPORT_AUDIO_STATUS = 0x7a,
+  CEC_ID_ROUTING_CHANGE = 0x80,
+  CEC_ID_ACTIVE_SOURCE = 0x82,
+  CEC_ID_GIVE_PHYSICAL_ADDRESS = 0x83,
+  CEC_ID_REPORT_PHYSICAL_ADDRESS = 0x84,
+  CEC_ID_REQUEST_ACTIVE_SOURCE = 0x85,
+  CEC_ID_SET_STREAM_PATH = 0x86,
+  CEC_ID_DEVICE_VENDOR_ID = 0x87,
+  CEC_ID_GIVE_DEVICE_VENDOR_ID = 0x8c,
+  CEC_ID_MENU_STATUS = 0x8e,
+  CEC_ID_GIVE_DEVICE_POWER_STATUS = 0x8f,
+  CEC_ID_REPORT_POWER_STATUS = 0x90,
+  CEC_ID_GET_MENU_LANGUAGE = 0x81,
+  CEC_ID_INACTIVE_SOURCE = 0x9d,
+  CEC_ID_CEC_VERSION = 0x9e,
+  CEC_ID_GET_CEC_VERSION = 0x9f,
+  CEC_ID_VENDOR_COMMAND_WITH_ID = 0xa0,
+  CEC_ID_REQUEST_ARC_INITIATION = 0xc3,
+  CEC_ID_ABORT = 0xff,
+} cec_id_t;
+
+typedef enum {
+  CEC_ABORT_UNRECOGNIZED = 0,
+  CEC_ABORT_INCORRECT_MODE = 1,
+  CEC_ABORT_NO_SOURCE = 2,
+  CEC_ABORT_INVALID = 3,
+  CEC_ABORT_REFUSED = 4,
+  CEC_ABORT_UNDETERMINED = 5,
+} cec_abort_t;
+
+#endif
